+++ packages/app/src/containers/ResultsPage/index.ts
@@ -1,0 +1,1 @@
+export { default as ResultsPage } from "./ResultsPage";
